# public/api/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ownHackathon API Overview",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/account/logout": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Log out the current user",
                "description": "Terminates the user session by invalidating the current access token. Clients should delete all locally stored tokens (Access Token and Refresh Token) after a successful logout.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Logout successful. The session has been invalidated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized. The access token is missing, expired, or already invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpResponseMessage"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "accessToken": []
                    }
                ]
            }
        },
...
        "/workspace": {
            "post": {
                "tags": [
                    "Workspace"
                ],
                "summary": "Create a new workspace",
                "description": "Creates a new workspace for the authenticated user. Workspaces are used to group related resources (e.g., events).",
                "operationId": "createWorkspace",
                "requestBody": {
                    "description": "The data for the new workspace. At a minimum, a name is required.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Workspace"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Workspace created successfully. The response contains the details of the newly created workspace.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Workspace"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input. The provided workspace name may be empty, too long, or contain invalid characters.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpResponseMessage"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication failed. The access token is missing or invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpResponseMessage"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "A workspace with this name already exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpResponseMessage"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "accessToken": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "AuthenticationResponse": {
                "properties": {
                    "accessToken": {
                        "description": "The access token after a valid log-in",
                        "type": "string"
                    },
                    "refreshToken": {
                        "description": "The refresh token after a valid log-in",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "HttpResponseMessage": {
                "properties": {
                    "statusCode": {
                        "description": "The Http Status Code",
                        "type": "integer",
                        "example": 400
                    },
                    "message": {
                        "description": "The Message",
                        "type": "string",
                        "example": "Bad request"
                    }
                },
                "type": "object"
            },
...
            "Workspace": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "The name from workspace",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "accessToken": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            },
            "refreshToken": {
                "type": "apiKey",
                "name": "Authentication",
                "in": "header"
            },
            "Client-Identification-String": {
                "type": "apiKey",
                "name": "x-ident",
                "in": "header"
            }
        }
    },
...
}